refactor: rename engine modules to snake_case
