+++ glacium/recipes/pointwise.py
@@ -1,7 +1,7 @@
 """Recipe integrating Pointwise mesh generation jobs."""
 
 from glacium.managers.RecipeManager import RecipeManager, BaseRecipe
-from glacium.engines.PointwiseJobs import PointwiseGCIJob, PointwiseMesh2Job
+from glacium.engines.pointwise_jobs import PointwiseGCIJob, PointwiseMesh2Job
 from glacium.engines.fluent2fensap import Fluent2FensapJob
 
 @RecipeManager.register
